The script to import program managers is now considering all existing academic years.
